requirements.txt
amqp==2.5.1
billiard==3.5.0.5
celery==4.2.1
Django==2.2.6
django-crispy-forms==1.8.0
django-debug-toolbar==2.0
django-extensions==2.2.3
django-filter==2.2.0
django-filters==0.2.1
djangorestframework==3.10.3
importlib-metadata==0.23
kombu==4.6.5
more-itertools==7.2.0
Pillow==6.2.0
python-memcached==1.59
pytz==2019.3
six==1.12.0
sqlparse==0.3.0
vine==1.3.0
zipp==0.6.0
